backend/requirements.txt
# Hotel Edge System - Backend
# Python FastAPI Backend

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Redis
redis==5.0.1
aioredis==2.0.1

# MQTT (IoT Device)
paho-mqtt==1.6.1

# Utilities
python-dateutil==2.8.2
httpx==0.26.0
python-dotenv==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
